Concurrent sweep: extract provider-api runtime to @agent-native/core; brain/dispatch/analytics adopt it
