+++ packages/core/package.json
@@ -82,6 +82,7 @@
     "./tracking": "./dist/tracking/index.js",
     "./usage": "./dist/usage/store.js",
     "./connections": "./dist/connections/index.js",
+    "./provider-api": "./dist/provider-api/index.js",
     "./code-agents": "./dist/code-agents/index.js",
     "./code-agents/transcript-normalizer": "./dist/code-agents/transcript-normalizer.js",
     "./workspace-connections": "./dist/workspace-connections/index.js",
